v0.23.8: fix(exc): subclass ``AttributeError`` for error raised in property [#8]
Reflection functions like ``hasattr`` access attributes when doing their
checks on objects. This is an issue when the attribute is a property,
since this will cause the code in the property's definition to be run.

The ``result`` attribute of ``StreamHMAC`` & subclasses is a property.
If their instances haven't been finalized, accessing ``result`` will
throw ``PermissionError``. This causes ``hasattr`` checks to throw
``PermissionError`` instead of returning a boolean. Asking ``hasattr``
of an object shouldn't raise an error. Creating a custom exception
that subclasses from ``AttributeError`` will cause ``hasattr`` to
respond correctly by returning a boolean.

Partial resolve [#8]