v0.23.10: fix(commons): deduplicate logic to keep behavior consistent [#14]
The ``__contains__`` logic also needed to be fixed to allow the
simplification in this commit. Checking ``hasattr`` wasn't a limiting
enough condition, as anything in the class dictionary would also
return ``True``. This led to errors when using namespaces as modules,
because they'd take in attribute names like ``__doc__`` which would
proc a ``PermissionError`` in frozen classes.

Resolution related to (#9, #11, #14)