build: fix improperly packaged directory structure for wheel
The wheel file was improperly packaged. This lead to pip installs
from the wheel to scatter the package's module's and subpackages
across the site-packages directory.

It's unclear if this change will undo the progress towards builds
which pass the tests on piwheels. However, it's more important to
not provide broken installs for pip installers than to pass those
third-party tests.