v0.23.16: build: send build data over sockets to a local signing service
The old package signing script was a hassle to run. The database needed
to be logged into with secure credentials for every signature. This was
a waste of time, a flow disrupter, was error prone, & didn't facilitate
frequent signatures.

This new design runs the package signer as a background service which is
logged into once, and subsequent requests to sign the package are sent
over sockets to the service. The integrity of the channel is enforced by
the knowledge of a transmission key, held as an environment variable
during local development. This allows the requesting script to run as
intended by only pressing enter on the default CLI options.

There were several options for making an easier signing flow. This
service oriented approach was chosen for its flexibility, with optional
support for decoupling of signing & development devices.

 _____________________________________
|                                     |
|            Usage Example:           |
|_____________________________________|

Terminal #1:

~$ source <path/to/secrets/env/file.sh>
~$ uv run sign_package_service.py  # CLI that awaits requests

Terminal #2:

~$ source <path/to/secrets/env/file.sh>
~$ uv run sign_package_build.py  # CLI to request a new signature