fix(test): collect names once if pytest runs w/o module specified
The modules used to aggregate tests based on their subdirectory
location within the package were also named with a prepended "test_".
This made it so that running ``pytest`` had different results than
running ``pytest tests/test_aiootp.py``, or another specified module.
Essentially, the tests were being collected multiple times, once for
every aggregator a test passed through while percolating up to the
namespace of the specified module. This caused internal pytest errors
& bugs.

The solution was to remove the "test_" prefix from aggregator modules.
The new names for aggregators begin with an "all_" prefix, which
should make it clear to all testers which modules to specify to test
only specific subsections of the package.