v0.23.18: fix(slots): refine splat logic, dir results & attr restrictions
For classes derived from `Slots`, there exist ways to decide which
instance variables are produced in `*obj` & `**obj` unpacking
interfaces, and in which order. Any instance variable not in the
`_UNMAPPED_ATTRIBUTES` frozenset, will be unpacked in the order they
are first declared: in the slots along the hierarchy's MRO; then,
in the __dict__ during an instance's lifetime. There was another
`_MAPPED_ATTRIBUTES` tuple which was removed for being unnecessary.

Additionally, any variable within `_UNMAPPED_ATTRIBUTES` was also
being omitted from the instance's `dir()` list. This made the logic
unweildy, because it had multiple purposes. So instead, the
`_DIRLESS_ATTRIBUTES` frozenset was added to specify which variables
should not appear in `dir()`.

Finally, a new `_RESTRICTED_ATTRIBUTES` frozenset was added as a
way to specify attributes which cannot be assigned or defined
as slots. This is convenient in general, but was necessary in
particular because allowing containers to override the non-private
`keys`, `values`, `items`, and `clear` methods with arbitrary
instance data would break the class' internal functionality.